fix: fix mismmatch of pointer types for slot in time based native scripts

+++ lib/src/scripts/native_scripts/script_invalid_after.c
@@ -248,7 +248,7 @@ cardano_script_invalid_after_from_json(const char* json, size_t json_size, carda
     return CARDANO_ERROR_INVALID_JSON;
   }
 
-  size_t slot = 0U;
+  uint64_t slot = 0U;
 
   cardano_error_t result = cardano_json_object_get_uint(slot_object, &slot);
